requirements/production.txt
asgiref==3.5.2
Django==4.1.4
django-environ==0.9.0
sqlparse==0.4.3
django-ordered-model
django-ckeditor
django-light
django-tailwind
django-ipware
loguru